Fixed non-unique offerId per problem in dev mode [#160796641]
Previously dev mode would use a constant offeringId of "1" which caused the section documents to be persistent across problems.

This change generates a unique offeringId based on the problem in the urlParams.

This change also adds a debug helper: in non-authed mode when you hover over the username you will now see the firebase uid.