Using ghost users more consistently and clarifying their behavior
Previously, `isGhostUser` was calculated in the Document component, but not passed to children/used as a prop in some places. Children now take and use that prop and are commented to explain its use.

`isGhostUser` is also marked optional in the Document component now, so it can be ignored when Documents are rendered in user-agnostic places like the Learning Log.