Fix drag of overlapped polygons [#162456045]
The code assumed that only one polygon would receive pointerDown/pointerUp messages. In fact, JSXGraph dispatches pointerDown/pointerUp messages to all objects at the affected location. If the "wrong" polygon received the events first, it would clear the intermediate drag cache so that the changed location of the dragged polygon would not be saved.