Add stamp button and menu to drawing toolbar
This created a button which will only be visible if stamps are defined.

The button shows both the current button and, nested, a small toggle to
open a fly-out menu of all the buttons. This is a combination of the
internal-twisty UI from the Workspaces drawing tool and the settings
menu.

If there is only one stamp, only the button is shown, with no toggle.

[#161492950]