Drag toolbar buttons to insert new tiles
This makes the toolbar buttons draggable, and gives them a new drag data
type of "tile.create" with a value of the tool type.

The document-content new accepts this data type for the purposes of
showing the blue insertion markers, and when dropped we use the new
`addTile` function from `document-content` which now accepts row info.

[#163273200]