Fix copy polygon behavior after selection changes made by annotation feature [#163922781]
Prior to the annotation feature polygons were not explicitly selected at the model layer; they were simply assumed to be selected if all of their vertices were selected. Code that assumed that polygons would not be among the selected objects must be updated to account for that possibility. In this case, the copy code ends up copying the same polygon twice -- once because it's now selected and once because all of its vertices are selected.