Updating the toolbar after a polygon is created
Tests were failing because selected points in a polygon were not enabling the angle label tool. There were two bugs involved in this: first, the selectedObjects getter was an action instead of a view so it wasn't observed and changes didn't re-render the toolbar. This was an easy fix.

The second problem was that just rendering when the selected point changes is not sufficient - when you double-click a point to create a polygon, the angle button should become enabled, even though the selected point hasn't changed. I don't believe that properties on the selected point are observed, so adding  the polygon didn't cause a re-render.  My quick fix was to pass down a function that can forceUpdate and cause a render when a polygon is created.

Of course, all these pieces worked before because the model was constantly re-created which forced re-renders.