Text tool working without rebound
Before we stored the entire TextContentModel in the state as prevContent and used that for comparison in getDerivedStateFromProps to determine if something had changed.  This model was causing trouble with the autorun function, as it was failing to fire.  Instead I stored just the text from the model in a private variable and used that for change comparison in the autorun - this allowed the autorun to successfully fire on text changes.
[#165457351]