Separate plot button from plot, add plot to nodes
Because of the way the UI is played out, the plot itself cannot be a
part of the plot button control component.

This extracts the plot from the plot button control, and adds in as
part of the node definition itself. The plot will only display when
we have a plot button control, so nodes such as the Data Storage will
never show a plot.