Set `programRunState` to `Running` in `runProgram()`
Previously, the `programRunState` could only be set to `Running` in `initProgramEditor()` which is called from `componentDidMount()`. This relied on the fact that switching from a ProblemDocument to a PersonalDocument required a complete reload of the component. Now that we're (correctly) using a PersonalDocument for the editable document, the `DataflowProgram` component is no longer unmounted/remounted, and so we must handle the state change internally.