New feature to copy documents into new documents from title bar.
Users can now create new documents that are copies of Problem,
Personal, and Learning Logs. Problem and Personal documents, when
copied, become a Personal document; Learning Logs, when copied,
become a Learning Log document.

Users are prompted for a new document title when copying a
document; document titles that are not renamed by a user will
default to "Copy of" preceding the original document title.