Ensure interval value remains in bounds
This adds a generic in-bounds checker for the drop-down value, which
attempts to pick the best value out of those remaining when some are
disabled.

In the case of the interval, it will pick the lowest interval or the
highest interval as appropriate, given the user's previous selection.

This also has two other effects:

1. ensuring that invalid values saved to a user's data are migrated to
   a valid value
2. enforcing the disabled list, as a user can no longer select a
   disabled value