Added desktop and data uri image drag/drop [#163658625]
Extracted image drag/drop handling from image-tool component to utility library so it could also be used in the document workspace component.

In addition to extracting the existing code new code was added to handle drop/drops of image files and image data uris.

To enable drops anywhere in the workspace styling was added to the workspace component so that it overlays the entire page.