Prevent deletion of last Personal or Learning Log Documents.
This fixes and prevents the issue of users from being able to delete
the last document of either Personal or Learning Log Documents, as
doing so would cause users to be unable to create additional documents
of either type.

The Delete button is set to the Delete Disabled button when documents
hit their last document of their type. Additionally as a fallback if
a user is able to accidentally access deleting the last document,
an alert pop-up will notify the user they have come across an error,
preventing the document from being deleted.