Fixed missing filter for latest publications in groups
The teacher dashboard was not showing the latest publication but rather the first publication found for each group member.  This caused the starred publication not to show in the right nav as it filters by latest publication.

Also refactors the document model to use a common function for getting the latest publication from a list of publications and removes use of "this" in two of the model views.