Removed wait for logging when switching classes or groups [#169345486]
This othen caused a long wait before the page was refreshed.

The first attempt to mitigate this was to use `navigator.sendBeacon` to send a fire and forget log message that isn't cancelled like an AJAX request is when the page unloads.  However we need to be able to set the content-type on the log request and Chrome has disabled setting the content-type for now with `sendBeacon`.