Floating text toolbar (#849)
[#174930211]

Implements a floating toolbar below the tile for the text tile. Toolbar is only visible/enabled when the text tile has focus. Toolbar scrolls with the document to stay pinned to bottom of tile. Toolbar is pinned to edge of window if bottom of tile extends off screen and is hidden if the entire tile is offscreen. Design changes are not included in this PR.

The core of the implementation is encapsulated in a custom hook which computes the location of the toolbar and updates the location as the document is scrolled or the tile is resized.