Draw tool floating toolbar [#174718210] (#852)
[ #174718210 ]

As with the text toolbar, this doesn't update the draw toolbar to the latest specs, it just turns the existing toolbar horizontal and makes it float. To do so, the existing code was extensively refactored into functional components using hooks. Taking the implementation from the text tool one step further, the tool's responsibilities are now encapsulated in the useToolbarToolApi() hook function.