Document context/types refactor (#854)
Document context/types refactor
- provide document context for thumbnails

One aspect of the teacher supports/image access solution requires that we know the context in which an image tile occurs -- image tiles in multi-class documents stored in firestore will require use of a cloud function while image tiles in other documents can be handled as they are now. Toward that end image tiles (and by extension all tiles) should have access to information about the context in which they occur. In fact, we already have a `DocumentContext` mechanism for providing this information in some cases, but it wasn't being used consistently, e.g. it was provided for editable documents only. With this PR we make the `DocumentContext` available more broadly.

Separately, there were a number of awkward module dependency relations with many low-level modules requiring access to high-level modules (notably `document.ts`) simply to gain access to some types. The solution here is to move a bunch of low-level types into a new `document-types.ts` module, and rearrange some of the remaining code so that the dependency relationships are more sensible, including the definition and use of the `DocumentContext`.

The result is a lot of changed files but the changes are minor in most of them.