Implement cloud function to support images in multi-class supports [#175363234] (#855)
Prior to the introduction of multi-class teacher supports, all CLUE content was scoped to the class in the firebase realtime database, including images. Multi-class teacher supports are stored in firestore and take advantage of firestore's more advanced security rules to enable users in any of a teacher's classes to access the support content. If a multi-class teacher support includes an image stored in the firebase realtime database, however, there's nothing those advanced firestore security rules can do to enable such multi-class access.

Rather than abandoning our image storage mechanism (again!) we implement a google cloud function which reads the multi-class support document from firebase, determines the classes that should have access to the document (and hence to its embedded images) and retrieves the desired image data if and only if the requesting user is in one of the classes that should be allowed access to the image.

Google cloud functions have admin-level access to firebase databases without being subject to any pesky security rules, so they must implement any appropriate security restrictions in code. As such, the `getImageData()` cloud function takes pains to verify that for authenticated users, claims retrieved from the firebase JWT match what is allowed according to the support document stored in firestore. For unauthenticated users (e.g. demo/dev/qa) we verify that what the client passes is appropriate, but since there's no guarantee that a client is following the rules, we at least make sure that unauthenticated users can only access parts of the database that don't require authentication.