Fix initial text toolbar location [#175631464] (#867)
The bug is essentially an order-of-initialization problem. During the initialization sequence the resize handler is called before the text toolbar has installed its handler, so it doesn't know the size of the tile when it comes time to determine the initial toolbar location. The fix here is to have the TextTool cache the size locally so that when the toolbar installs its handler the handler can be provided with the cached dimensions.