Table UI IV: Implement Add Column and Remove Row buttons [#175773902] (#872)
I was originally just going to implement Remove Row in this PR, but it turns out that supporting Add Column as well didn't add much to the degree of difficulty.

Note that for implementation reasons, the Add Column and Remove Row buttons are implemented as an additional column in the table. This makes rendering them straightforward using normal React mechanisms, guarantees proper alignment with the row, assures that they scroll properly when the table is scrolled vertically, etc. The downside to this approach, however, is that if the table is wider than the tile the buttons can be scrolled out of view horizontally. In the interest of moving on I propose we stick with this for now. If there is a desire to implement sticky buttons that can't be scrolled out of view, we should make another story since the implementation will be non-trivial.

On another note, when new columns are created they are given the first available name in the sequence "y", "y2", "y3", etc. I experimented with using subscripts as suggested by the spec but in the absence of rich-text editing we're limited to Unicode subscripts which by default render too small and are difficult to edit/enter for the user, so I decided to skip it.