Re-implement SingleStringDialog in terms of new useCustomModal() (#875)
As a further test of the new custom dialog infrastructure, we replace the existing <SingleStringDialog> component with an API-compatible version based on the new useCustomModal hook. One wrinkle is that the only existing client of the <SingleStringDialog> (the comment feature in the geometry tool) is a class component, which therefore can't use the useSingleStringDialog hook. Not to be deterred, however, with this PR we also implement a replacement <SingleStringDialog> as a functional component which wraps the useSingleStringDialog hook.