Fix error message feedback in useSetExpressionDialog()
This required a larger refactor than one might suspect. The previous code treated the expression input as an uncontrolled component, i.e. no call to `setState()` was required for individual keystrokes. The error message has to be a controlled component since it's only set from code, thus setting the error message requires a `setState()` call, which triggers a re-render of the component. When that occurred, the expression input would lose focus, which could occur after each keystroke. The problem is that the previous code was regenerating a new instance of the `Content` functional component whenever any of its inputs changed because it closed over its required values rather than taking them as props. With this change, the `useCustomModal()` code now supports/encourages the use of static function components that take props which can be passed separately, which allows controlled components to behave as expected. This is a much better model moving forward.