Fix table height and graph/table automatic title generation in curriculum content (#913)
- We were only enabling the ToolApiInterface in editable documents, but for title and height determinations it needs to be extended to read-only document contexts as well. Therefore, we move responsibility for the ToolApiInterface to the CanvasComponent and use React context to distribute it to clients rather than prop drilling.

This PR also implements automatic sizing of curriculum rows to the expected size of tables. (This approach can be extended to other tiles, but only tables support it for now.) The manual height set in the content can still override, however, e.g. for very long tables where the user is expected to scroll. Thus, to enable the automatic sizing I removed all of the explicit heights from the curriculum tables in Moving Straight Ahead, as they all seemed to warrant automatic sizing.

This PR also enables column width resizing in tables in curriculum content, which were originally disabled as part of supporting the readOnly flag, but column resizing can still be useful in read-only tables.

Finally, this PR also fixes a bug in table title cell sizing in read-only content.

These changes are more extensive than I'd like given how close we are to release, but most of the changes are mechanical, i.e. pull the ToolApiInterface from context rather than from props, so the impact is less than it appears at first blush.