Archive code to export student groups [#172209086] (#935)
Co-authored-by: mklewandowski (who did all of the original work)

This component was developed to generate a CSV-formatted report for teachers of the student groupings for each problem (offering) in a class. This was required at the time because early versions of CLUE did not log group information and so there was no easy way to determine student groupings from the existing logs. A student's `groupId` was added to the existing logs for CLUE 0.8.0 in [PR #451](https://github.com/concord-consortium/collaborative-learning/pull/451), obviating the need for this particular report component and its button. This component is left in the code base as an exemplar in case future need for CSV export of CLUE data arises. [[#172209086]](https://www.pivotaltracker.com/story/show/172209086)