Make `safeJsonParse()` type-safe (#937)
In working on the export capability for geometry tiles, I am frequently working with the changes array, which involves parsing JSON (because the changes are `JSON.stringify`ed). `JSON.parse()` is inherently not type-safe, as its return value is `any` and it throws an exception if it encounters anything it doesn't recognize. We have our own `safeJsonParse()` utility function which converts the exception into an `undefined` return value but doesn't do anything to address the lack of type-safety. A common usage pattern for `safeJsonParse()` looks something like:
```typescript
  const foo = safeJsonParse(fooJson) as Foo;
```

This has the advantage of providing type information immediately, but unfortunately it loses the fact that `undefined` is a possible return value. Thus,
```typescript
  const foo = safeJsonParse(fooJson) as Foo | undefined;
```
would be more accurate, but it's easy to forget that, which just indicates that type-safe use of `safeJsonParse()` is error-prone at best and non-existent at worst.

With this PR we add a generic type parameter to `safeJsonParse()` so that the preceding can be written as:
```typescript
  const foo = safeJsonParse<Foo>(fooJson);
```
which maintains proper type-safety. Further, this PR also visits most uses of `safeJsonParse()` to update to the type-safe usage. In many cases this is just a syntactic change with no other code change required, but in others the possibility of an `undefined` return value was not being handled correctly and so minor code changes were required to account for that possibility.