Fix table/graph title incrementing [#177273772] (#939)
The unique titling is mediated by the document's `ToolApiInterface`, which is made available to clients via React context. In the original implementation, tiles were created without titles and then were given dynamically generated titles when instantiated. If users didn't explicitly edit the title, however, this dynamically generated title was not saved, which meant that the same tile might get a different dynamically generated title the next time the same document was opened.

Recently, this mechanism was enhanced so that newly created tiles would get a persistent unique title on creation (in theory) rather than waiting for it to be dynamically generated at instantiation time. As a result, a given tile would be able to retain its unique title on every opening of the document. Unfortunately, the Toolbar (which provides the UI for the creation of new tiles) was not in a part of the React component tree that has access to the `ToolApiInterface` context, which resulted in the failure of the unique titling. 🤦

The fix here is to create another piece of React context that encompasses more of the React component tree and which allows the Canvas (which creates the `ToolApiInterface`) to set it so that the Toolbar can access it, which fixes the unique titling on tile creation.

Looking at the history, earlier the `ToolApiInterface` was only available in editable documents. This meant that tiles in curriculum content weren't given unique titles. To fix that, we recently made the change to create the `ToolApiInterface` in the Canvas and distribute it to all tiles (editable or not) via React context. It was in this transition that we inadvertently broke unique titling for editable documents due to the Toolbar not having access to the necessary context.