174765829 graph link to table (#942)
Adds ability to link table from graph via a link icon.
Clicking on the button opens a modal dialog that lets user choose tables to link or unlink.
Button is enabled when there is a table present in the document, regardless of the table content, and can be selected from the Link/Unlink dialog. Error dialog does appear if selected table has invalid content.