Add logging of monitoring/unmonitoring user documents (#943)
This is primarily for debugging purposes. We have had two recent reports of users being able to work locally without incident while their efforts are not saved to firebase. The mechanism by which this is supposed to happen is that we monitor the user's document such that any changes trigger a save to firebase. The reported symptoms are consistent with a failure of this monitoring mechanism, so we log it in hopes of learning something.