Temporary workaround for images copied from multi-class supports (#944)
This allows images copied from multi-class teacher supports to user documents to load successfully rather than showing the loading spinner indefinitely. I consider this to be a workaround rather than a proper fix because it doesn't prevent the initial error from occurring, which results in a console error and a Rollbar error report. It merely patches over the error from the user's perspective after the fact. But in the near term, it's an improvement in the user experience and so seems worth it until we find time to address the underlying issue.