Log the document change count with log messages for use in debugging (#947)
While the 1.8.0 release is currently held up due to deploy issues, it occurred to me that logging the document `changeCount` could be useful. The `changeCount` is incremented by the same listener that is responsible for triggering a save to firebase, so we may be able to use the logging of events that should increment the `changeCount` to identify situations in which document changes aren't triggering a save.