Fix parse error in `prettifyExpression()` with column names with spaces (#959)
The `prettifyExpression()` function formats a table column expression for display to the user. Under the hood it was using the `substitute()` function of the `expr-eval` library we're using for expression evaluation to insert the name of any columns referenced. The `substitute()` function expects the result of the expression to be parseable/evaluable, which is not the case when the referenced column names contain spaces. But since `prettifyExpression()` is only intended to format expressions for display to the user, a simple string replacement is sufficient for our needs.