178762390 panel divider (#996)
Adds the collapsed tabs for My Resources, and My Workspaces, and the resize bar with thumbnail.

My Workspaces is expanded by default. 
Clicking on the collapsed My Resources or My Workspaces tabs expand them. Currently, My Resources only expands to half the workspace, while My Workspaces expand to full or half width of the workspace.
To collapse My Resources, click on the "x" close icon on the Resource Panel.
To collapse My Workspaces, click on the resize thumbnail.