disable chromeWebSecurity and clearCookies on each login
chromeWebSecurity had to be disabled due to an error we could not full track down.
When running headless with electron a second test would fail with a cross domain error.
This only happened if the first test was run before the second test.
And in all there were about 6 navigations and only the last navigation of the last test was
causing an error.
All of the logging we added indicated there was no cross domain navigation happening.
We recorded the url after each navigation and with web security disabled none of the
these urls changed the "superdomain".

The clearCookies change might not be necessary with this use of the custom login
command, but it was needed in the activity player when the cypress test runner was left
open for longer than the portal session timeout.