remove waitForSpinner, improve qaClear, and use baseUrl
waitForSpinner was inherently error prone.
If the app loaded slowly the spinner would not be shown yet so the test for the spinner would pass when
the intention was for the test to only pass once the spinner was visible and then disappeared.
This meant that a wait of 2s was added to give the spinner time to show up, but this slows down all of the
tests. And occasionally it would take longer than 2s for the spinner to show up then the test could fail.

qaClear used waitForSpinner, it is now optimized to just wait directly for the 'QA Cleared: OK'.

cypress will automatically make all cy.visit urls relative to the baseUrl so there isn't a need to
look up the baseUrl and concatenate the URLs.