Add ToolComponent to ToolContentInfo
To prevent circular dependency issues this also required splitting the tool registration
out from content models into their own files, and moving the registration out of
the tool-tile content module.

Because the registration doesn't happen in a module that is loaded by everything,
now certain tests need to do an explicit registration in order to pass.