a test failed because text was not saved before the page navigation
This adds a wait after typing the text.
There doesn't seem to be any debouncing of the text so it should be saved quickly.
In most cases this seemed to be working, hopefully the wait will decrease the failures.
As the comment says, ideally there would be a wait make sure the data was actually
saved.