pass the actual ToolButtonModel objects to tool-button.tsx
This is a first step towards removing the use of the DocumentTool type
in the toolbar. Instead of having a enumeration of strings, most of the
toolbar functions can take a ToolButtonModel.

It might also be possible to remove most of these functions and make
them actions on the ToolButtonModel instead.