Switch to passing around tool button model
This simplifies a little bit of code because the Icon property can now just be taken
from the model.

It also adds a bit more typesafety because the DocumentTool type needs to be
removed to support dynamic tools.

It does add 2 casts of strings to DocumentTool to use the existing
IDragToolCreateInfo interface and addTile action.