switch DocumentTool to string
now that tools added by calling the registerToolContentInfo method,
we cannot use static typing with the tool ids.

I'd also note that the places where the tool value is passed to the changed
methods it was just being casted from a string. So there wan't much type
safety here.

It was nice to know the type of the argument should be a tool id. And it would
be possible to just alias string as DocumentTool. But doing this kind of
creates a false sense of security. Typescript would take a string or a
DocumentTool in this case it would not even require it to be casted.
Because of this false security it seems better to just use string.