code to support undo, change documents, and shared models
this isn't meant to be merged, it is just a demo
with this commit the undo only works works with the geometry tool
there is a separate PR which modifies the other tiles so they work
with undo.
there isn't an example of the shared model here, and more code
needs to be added to support them.