add the generic type to the content property
Previously the type of this content property was 'any'.
This any type hid several issues.

Now the type is the generic ToolContentModel
Doing this means we can add additional fields to that generic model
and the CLUE code can assume they exist.

I tested and if a tile doesn't 'extend' ToolContentModel and is missing
required properties there are several typescript errors that show up to
catch this case.
Additionally we can use tool-tile.test.ts  to verify that each built in tile we
add supports these properties.

Adding this typing turned up a few hidden issues:
- TextTool wasn't following extension pattern
- common checking with `toolType && getToolContentInfoById(toolType)` wasn't
safe because toolType could be the empty string.  This was masked because toolType
was 'any' in some cases instead of 'string'.