first pass on shared variables in text tile
The dialog that works in the slate-editor wasn't easy to add to clue, so right now it just
picks up the first shared variable when one is added.

This changes things to use a 'chip' component that responds to changes in the shared
variable.

Note this has a type error that will be fixed shortly.