Handle deletion of shared variables
This does something that might not be great.
It adds the editor to the content model.
This way changes in the content model can directly affect the editor.
Normally a view would respond to the content model but in this case the
editor doesn't.
Additionally the editor provides all of the methods to modify the content, so
it isn't practical to move all of that into the content model itself.

This also illustrates another feature that text plugins need to support:
customized updateAfterSharedModelChanges handling.