switch to using the document as the tree root
Previously the DocumentsModel was the tree root

The main impact on CLUE is that the array of documents in DocumentsModel
is now a volatile prop instead part of the state.
The DocumentsModel was not being serialized anyhow so I don't think this makes
much of a difference.

Undo is no longer working, but I suspect it was broken before these changes.

This fixes the issue with failure to load multiple documents that had duplicate ids.