fix chip selection behavior
the key thing was fixing the schema.
this also fixes up some other behaviors around replacing the existing node.

Note there is still an issue when a chip is selected and then v= tool is clicked the variable edit dialog doesn't show the current variable, and any changes do not replace the selected chip.