each document is its own MST tree
This will help when adding middleware that is specific to a document.
With this change, each documents middleware can record just the changes in that document.

Additionally if references are used and a document is copied, these references will still be vaild.
When there is just one tree all ids have to unique to the tree so a document and its copy will have conflicts unless the ids are changed during the copy.