remove cypress global before and after hooks
When running multiple specs these hooks all get run at the beginning of all of the specs
So when running the whole suite the qa clear would happen 12 times, and several
pages would be visited at the wrong time.

I suspect this has been the source of some of the CI flakey-ness. I don't know for sure
but I'd guess the test runner runs multiple specs at the same time.