bring in the latest diagram view library
To use this latest library it is necessary to have some basic shared variable support.
The state saved by the diagram tile using this interim shared variable support is not
compatible with the old diagram tile. And it also won't be compatible when the real
shared variable support is added.

TODO:
- figure out what to do with the incompatibility
- add test coverage