save shared model by label in tool Tile
I looked at letting the tile content model just save the shared model reference
in its own state. This way the content model could organize it however it wanted.
But we also need the container to know which shared models a tile references.
MST supports walking the tree, but this seems more difficult than just storing
the references in the tool tile model so they are all in one place.

TODO:
- need new tests for shared model framework
- need to move monitoring of shared model changes to document or tree
- try to build the initial update call into the framework