add indirection to fix MST late problem
The issue was that types.map(types.late(...)), was causing the late to evaluate
immediately.  The only reasonable solution I could find for this was to add an
intermediate model that was the actual map value.

Other options that didn't work out:
types.late(types.map(types.late(...))) did not work because it didn't provide a default
value for the map. So any place loading or creating the model would have to change
to support passing `{}` for the late map field.

types.optional(types.late(types.map(types.late(...))), {}) didn't work. This triggered the
first late to be evaluated immediately, probably this is to figure out how to process the
default value of {}.

types.maybe(types.late(types.map(type.late(...)))) didn't work because there isn't a way
that I could find to create the map at runtime. It would start out undefined because of
the maybe and then I'd need a way to do MapType.create(). But this isn't documented
or supported.