remove duplicate monitoring of shared models
The autorun that the SharedModelDocumentManager was starting
isn't necessary when the tree monitor is present.

Change the path the tree monitor was monitoring since the
shared models are not inside of shared model entries.

Disable some Jest tests that now require the tree monitor to be
setup. Probably the best approach is to add these tests to document.test.ts
and perhaps move the test shared model and test tile definitions out of
shared-model-document-manager.test.ts and into a common file.