add variable chip support to the text tile
This is basically a rebase of work that was original done on top of the add-tree-monitor branch.

This code passes the Jest tests, but when running the local server documents aren't loading.

Even if that issue is fixed it shouldn't be merged because it hardcodes a dependency on the
SharedVariables shared model into the text tile. This should be replaced with a text plugin
mechanism so the SharedVariables shared model can register this text plugin.