fix issue when chip was edited
I was adding cypress tests and then ran into this issue.
It required a strange combination of steps to be exposed that I still don't understand.
The cypress test does these steps.

The visible effect was that editing a chip would cause other chips to disappear.
And in the console there'd be a warning about duplicate ids.
My guess is that the old code was creating a new slate node to replace the selected one,
so when this new node was added it delete the old node from the slate value.
However we still had a reference to this old node, so there was some strange interaction.
But it really doesn't make a lot of sense. If we run into something like this after slate is
updated I would dig through the slate code to try to figure out what is going on.