fix drawing stamp tool
When the document model was made the root of the MST tree instead of
the documents model, the stamps configuration stopped work.
This fixes that by adding the appConfig to the MST environment.
It does this by setting this in the environment when the document is added
to the documents list.

This also adds the stamp, stroke-color, and fill-color tools back to the
list of default tools.

[#182287732]