182265390 dataflow plugin (#1296)
* Bring in Dataflow components from dataflow branch

This is mostly a giant dump of all the components and models from the
dataflow branch of the project.

The majority of the content is brought over as-is, modulo changes to the
folder structure to match the current CLUE setup.

The big missing pieces are generally marked with `FIXME` in the code,
and are mostly confined to the IOT work in dataflow-program.tsx. No
attempt is made in this PR to run programs in AWS.

All the control components (e.g. `delete-control.tsx`) also need some
refactoring to appease the rule-of-hooks, but I did not figure out the
correct way to update them all.

Co-authored-by: Sam Fentress <sfentress@concord.org>
Co-authored-by: Kirk Swenson <kswenson@concord.org>