add withoutUndo
this replaces the MST utitlity actionTrackingMiddleware2 with
actionTrackingMiddleware3
This 3rd version pases the actual MST action call to the hooks.
This way we can record the action tracking call in a weak map.
This weak map is used to look up this action tracking call so it can be
modified by withoutUndo.