fix reusing canvas component, and teacher docs
The same canvas component is used for multiple documents,
so we can't just store the props.document when it is first created.
Instead on each render we look to see which document
we should be using the props, or the historyDocument.

With the change to use the document's uid when finding the metadata path,
this broke teacher documents which need to use the current teacher's info
Now the loading component knows if it is loaded from the teacher dashboard.
This way it only uses the document's uid when it is is loaded this way.

This commit also fixes:
Some test helpers which were looking at sections.
They were looking for sections anywhere,
but really they should just be in the primary workspace.

Now that we