build once, test multiple times, use cypress image
splits the build step into its own job, it is based on the doc here:
https://docs.cypress.io/guides/continuous-integration/github-actions
I think the use of the cypress action to do the building will also
enable better caching.
The use of a cypress image could allow us to go back to chrome.
The image might also let us test locally on the same image.