remove contextId
This property was not be read anywhere.
In firestore the context_id is used instead.
context_id is written by the firebase function from the request context.

The only sketchy thing was the logic in use-stores.ts.
It was setting the value from either the remoteContext or user object.
But since contextId is never read this still seemed save to remove.

Now document.ts can construct the IDocumentMetadata by itself.

The useMemo isn't needed anymore.
MST will return the same metadata object from get metadata()