add a main document to the TreeManager
This better represents how the tree manager is current working.
It is writing history entries under the main document key.
The IMainDocument interface is used to prevent a circular dependency.

In the future if we can separate the serialization code,
then this mainDocument could be moved to that separate module.