add DEPLOY_PATH support
Building an index-top.html that has a publicPath of DEPLOY_PATH, is required to support releasing with the s3 deploy github action.

This PR sets up that support and fixes places where assets are loaded relative to the index.html so they will instead be loaded relative to the public path.