update destroy model tests
This still needs an update to the documentation on this, describing the benefits and downsides of detach approach.
And it would be nice to add a test that used real react components.

With a real react setup, we might be able to demonstrate a more general approach.
Something like checkMSTModel(model) which just calls isAlive on the model and prints a warning if it is not alive. If this was the first function called in a component's render method it ought to short circuit the computedValue checks.